Modified the PlaylistManager/Service to have references to eachother for easier communication
