+++ library/build.gradle
@@ -2,8 +2,8 @@ apply plugin: 'com.android.library'
 apply plugin: 'maven'
 apply plugin: 'bintray-release'
 
-def versionMajor = 1
-def versionMinor = 1
+def versionMajor = 2
+def versionMinor = 0
 def versionPatch = 0
 
 // Maven GAV
@@ -56,7 +56,7 @@ uploadArchives {
             repository (url: "file://${System.env.HOME}/.m2/repository/")
             pom.groupId = libraryGroupId
             pom.artifactId = libraryBaseName
-            pom.version = libraryVersion
+            pom.version = libraryVersion + "-SNAPSHOT"
             pom.packaging = 'aar'
             uniqueVersion = false // don't put in a timestamp on the filename (other project will not be able to depend on it)
         }
