+++ library/src/main/java/com/devbrackets/android/exomedia/service/EMPlaylistService.java
@@ -233,6 +233,7 @@ public abstract class EMPlaylistService<I extends EMPlaylistManager.PlaylistItem
 
         notificationHelper = new EMNotification(getApplicationContext());
         lockScreenHelper = new EMLockScreen(getApplicationContext(), getClass());
+        getMediaPlaylistManager().registerService(this);
 
         //Another part of the workaround for some Samsung devices
         if (workaroundIntent != null) {
@@ -249,6 +250,7 @@ public abstract class EMPlaylistService<I extends EMPlaylistManager.PlaylistItem
         setMediaState(MediaState.STOPPED);
 
         relaxResources(true);
+        getMediaPlaylistManager().unRegisterService();
         audioFocusHelper.abandonFocus();
         lockScreenHelper.release();
